agents_app/requirements.txt
langchain>=0.0.267
langchain-openai>=0.0.2
pydantic>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
openai-agents